Critical Rendering Path (critical css): the Slider
The slider looks nice, but it takes way too long to load and there is no relevant/usable content in the meantime. If, for instance, the two buttons in the slider are loaded firsthand, it already gives the user the option to interact with the website instead of waiting 30 seconds.